+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+0.0.2 (2016-08-02)
+~~~~~~~~~~~~~~~~~~
+
+- Fixed bug with command line arguments
+- Added validate command
+
 0.0.1 (2016-07-29)
 ~~~~~~~~~~~~~~~~~~